@@ -2,6 +2,7 @@
   "name": "portfolio-web",
   "version": "2.2.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "prisma generate && next build",
@@ -9,34 +10,34 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@mantine/hooks": "^6.0.11",
-    "@prisma/client": "^4.14.1",
-    "@radix-ui/react-separator": "^1.0.2",
-    "@radix-ui/react-tooltip": "^1.0.5",
-    "@types/node": "20.2.3",
-    "@types/react": "18.2.6",
-    "@types/react-dom": "18.2.4",
-    "@vercel/analytics": "^1.0.1",
-    "autoprefixer": "10.4.14",
-    "class-variance-authority": "^0.6.0",
-    "clsx": "^1.2.1",
-    "emoji-regex": "^10.2.1",
-    "eslint": "8.41.0",
-    "eslint-config-next": "13.4.3",
-    "lucide-react": "^0.220.0",
-    "next": "13.4.3",
-    "next-cloudinary": "^4.10.0",
-    "postcss": "8.4.23",
+    "@mantine/hooks": "^7.4.2",
+    "@prisma/client": "^5.8.1",
+    "@radix-ui/react-separator": "^1.0.3",
+    "@radix-ui/react-tooltip": "^1.0.7",
+    "@types/node": "20.11.6",
+    "@types/react": "18.2.48",
+    "@types/react-dom": "18.2.18",
+    "@vercel/analytics": "^1.1.2",
+    "autoprefixer": "10.4.17",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.0",
+    "emoji-regex": "^10.3.0",
+    "eslint": "8.56.0",
+    "eslint-config-next": "14.1.0",
+    "lucide-react": "^0.314.0",
+    "next": "^14.1.0",
+    "next-cloudinary": "^5.20.0",
+    "postcss": "8.4.33",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-scroll-to-top": "^3.0.0",
-    "tailwind-merge": "^1.12.0",
-    "tailwindcss": "3.3.2",
-    "tailwindcss-animate": "^1.0.5",
-    "typescript": "5.0.4"
+    "tailwind-merge": "^2.2.1",
+    "tailwindcss": "3.4.1",
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "5.3.3"
   },
   "devDependencies": {
-    "@total-typescript/ts-reset": "^0.4.2",
-    "prisma": "^4.14.1"
+    "@total-typescript/ts-reset": "^0.5.1",
+    "prisma": "^5.8.1"
   }
 }
